% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCentromereSize.R
\name{CalculateCentromereSize}
\alias{CalculateCentromereSize}
\title{Calculate Centromere Size}
\usage{
CalculateCentromereSize(chr_total, centromere_pos_total)
}
\description{
This function Calculate Centromere Size
}
