% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDeletionTable.R
\name{CreateDeletionTable}
\alias{CreateDeletionTable}
\title{Create Deletion Table}
\usage{
CreateDeletionTable(vcf_table, chr_data)
}
\arguments{
\item{vcf_table}{}
}
\description{
This function returns the deletion table
}
