% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDupDelTable.R
\name{CreateDupDelTable}
\alias{CreateDupDelTable}
\title{Create Dup Del Table}
\usage{
CreateDupDelTable(window, accession, vcf_path, output_path, isPlot = TRUE)
}
\arguments{
\item{output_path}{}

\item{isPlot}{True or False, default is True}
}
\description{
Create Dup Del Table
}
