% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDuplicationTable.R
\name{CreateDuplicationTable}
\alias{CreateDuplicationTable}
\title{Create Duplication Table}
\usage{
CreateDuplicationTable(window, vcf_table, chr_data)
}
\arguments{
\item{window}{}

\item{vcf_table}{}
}
\description{
This function returns the duplication table
}
