% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVcfTable.R
\name{CreateVcfTable}
\alias{CreateVcfTable}
\title{CreateVcfTable}
\usage{
CreateVcfTable(accession, vcf_path, chr_data)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
The square of the input number.
}
\description{
This function returns the square of a number.
}
