% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatVCF.R
\name{CreatVCF}
\alias{CreatVCF}
\title{CreatVCF}
\usage{
CreatVCF(Directory, Genome_Fa, Picard_Path, GATK_Path)
}
\arguments{
\item{Directory}{The Path of to the BAM Directory}

\item{Genome_Fa}{Path for whole genome FASTQ file}

\item{Picard_Path}{Path to the Picard directory, with all the JAR files}

\item{GATK_Path}{Path to the GATK directory, where the GenomeAnalysisTK.jar file is located}
}
\value{
None
}
\description{
Create the VCF (Variant Call Format) file
}

