% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeletionTable.R
\name{DeletionTable}
\alias{DeletionTable}
\title{DeletionTable}
\usage{
DeletionTable(Directory, Table, dbSNP_Data_Directory, dbSNP_File_Name,
  Genome_Fa_dict, Organism)
}
\arguments{
\item{Directory}{The Path of to the BAM Directory, containing the variantTable.csv file}

\item{Table}{The variable containing the output of the MajorMinorCalc function}

\item{dbSNP_Data_Directory}{The path for the directory where the edited dbSNP file are (created previously by the Edit_dbSNP_Files function)}

\item{dbSNP_File_Name}{The edited dbSNP file names, without the chromosome number}

\item{Genome_Fa_dict}{the path for the dictionary file created for the whole genome FASTA file.}

\item{Organism}{"Human" or "Mouse"}
}
\value{
None
}
\description{
Intersect the observed SNPs with the common SNPs table from dbSNPs, Creates file with the LOH data called Deletions.txt
}

