% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edit_dbSNP_Files.R
\name{Edit_dbSNP_Files}
\alias{Edit_dbSNP_Files}
\title{Edit_dbSNP_Files}
\usage{
Edit_dbSNP_Files(Directory, File_Name, Organism)
}
\arguments{
\item{Directory}{the directory were the files are, one GTF file per chromosome}

\item{File_Name}{the files name, without the number of the chromosomes}

\item{Organism}{"Human" or "Mouse"}
}
\value{
None
}
\description{
Before LOH analysis, the dbSNP files need to be edited using the Edit_dbSNP_Files function (done only once):
}

