% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditVCF.R
\name{EditVCF}
\alias{EditVCF}
\title{EditVCF}
\usage{
EditVCF(Directory, Organism)
}
\arguments{
\item{Directory}{The Path of to the BAM Directory}

\item{Organism}{"Human" or "Mouse"}
}
\value{
None
}
\description{
Edit the VCF (Variant Call Format) file, Creates file with SNPs data at the BAM directory called variantTable.csv
}

