% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MajorMinorCalc.R
\name{MajorMinorCalc}
\alias{MajorMinorCalc}
\title{MajorMinorCalc}
\usage{
MajorMinorCalc(Table, minDP, maxDP, minAF)
}
\arguments{
\item{Table}{The variable containing the table of SNPs}

\item{minDP}{minimal reading depth required from accepted SNP, usually set to 20}

\item{maxDP}{maximal reading depth required from accepted SNP, usually set to very high number}

\item{minAF}{minimal minor allele frequency, usually set to 0.2-0.25}
}
\value{
None
}
\description{
Major to Minor Calculation function
}

