% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenSRA.R
\name{OpenSRA}
\alias{OpenSRA}
\title{OpenSRA}
\usage{
OpenSRA(File, Library_Type, SRAPath)
}
\arguments{
\item{File}{The Path to the SRA File}

\item{Library_Type}{"Paired" or "Single"}

\item{SRAPath}{path for the SRAtools BIN directory (example: SRAPath = "~/SRAtoolkit/sratoolkit.2.5.4-ubuntu64/bin/")}
}
\value{
None
}
\description{
Open SRA file into FASTQ file. The script will create new directory called FASTQ with the fastq files.
}

