% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Zygosity_Blocks.R
\name{Plot_Zygosity_Blocks}
\alias{Plot_Zygosity_Blocks}
\title{Plot_Zygosity_Blocks}
\usage{
Plot_Zygosity_Blocks(Table, Window, Max, Max2, Organism)
}
\arguments{
\item{Table}{The deletion table containing the output of the DeletionTable function}

\item{Max}{How many Heterozygouse SNP need to be in a block to get the full color, usually 6}

\item{Organism}{"Human" or "Mouse"}

\item{window}{the block size in bp, usually 1500000}

\item{Max}{How many Homozygouse SNP need to be in a block to get the full color, usually 60}
}
\value{
None
}
\description{
Plot blocks of heterozygous and homozygous SNPs
}

