% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Zygosity_Sinle.R
\name{Plot_Zygosity_Sinle}
\alias{Plot_Zygosity_Sinle}
\title{Plot_Zygosity_Sinle}
\usage{
Plot_Zygosity_Sinle(Table, Organism)
}
\arguments{
\item{Table}{The LOH table containing the output of the DeletionTable function}

\item{Organism}{"Human" or "Mouse"}
}
\value{
None
}
\description{
Plot each SNP, without any summarization
}

