% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGenome.R
\name{PlotGenome}
\alias{PlotGenome}
\title{PlotGenome}
\usage{
PlotGenome(orderedTable, Window, Ylim, PValue, Organism)
}
\arguments{
\item{orderedTable}{The variable containing the output of the MajorMinorCalc function}

\item{Window}{an odd number determining the window of the moving average plot, usually 151}

\item{Ylim}{the plot y axes maximal limit, usually 3}

\item{PValue}{Determine if to add the P-Value bar to the graph, accepts "TRUE" or "FALSE" values}

\item{Organism}{"Human" or "Mouse"}
}
\value{
None
}
\description{
Plot Allelic ratio along the genome for duplication detection.
}

