% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sort_major_minor.R
\name{Sort_major_minor}
\alias{Sort_major_minor}
\title{Sort_major_minor}
\usage{
Sort_major_minor(data, col1, col2)
}
\arguments{
\item{data}{data}

\item{col1}{col1}

\item{col2}{col2}
}
\value{
None
}
\description{
Sorting and reordering table
}

