% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRAtoSNP.R
\name{SRAtoSNP}
\alias{SRAtoSNP}
\title{SRAtoSNP}
\usage{
SRAtoSNP(File, Library_Type, TopHat_Threads, Transcript_Anotations,
  Bowtie_Genome_Index, Genome_FA, Picard_Path, GATK_Path, Organism)
}
\arguments{
\item{File}{The Path to the SRA File}

\item{Library_Type}{"Paired" or "Single"}

\item{TopHat_Threads}{Number of threads used for the alignment}

\item{Bowtie_Genome_Index}{The basename of the genome index to be searched. The basename is the name of any of the index files up to but not including the first period.}

\item{Picard_Path}{Path to the Picard directory, with all the JAR files}

\item{GATK_Path}{Path to the GATK directory, with GenomeAnalysisTK.jar file}

\item{Organism}{"Human" or "Mouse"}

\item{Transcripts_Annotations}{path for the transcript annotation GTF file.}

\item{Genome_Fa}{Path for whole genome FASTA file}

\item{dbSNP_Data_Directory}{The path for the directory where the edited dbSNP file are (created previously by the Edit_dbSNP_Files function)}

\item{dbSNP_File_Name}{The edited dbSNP file names, without the chromosome number}

\item{Genome_Fa_dict}{the path for the dictionary file created for the whole genome FASTA file.}
}
\value{
None
}
\description{
In case of analyzing SRA file, it is also possible to use the following function that performs step 1-4 in one function
}

