% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tophat.R
\name{Tophat}
\alias{Tophat}
\title{Tophat}
\usage{
Tophat(Directory, Library_Type, Threads, Transcripts_Annotations,
  Bowtie_Genome_Index)
}
\arguments{
\item{Directory}{The Path of the directory were the FASTQ files are}

\item{Library_Type}{"Paired" or "Single"}

\item{Threads}{Number of threads used for the alignment}

\item{Transcripts_Annotations}{path for the transcript annotation GTF file.}

\item{Bowtie_Genome_Index}{The basename of the genome index to be searched. The basename is the name of any of the index files up to but not including the first period.}
}
\value{
None
}
\description{
Align reads to genome using Tophat2. The Script creates new directory called BAM in the same path were the FASTQ dir is located.
}

